migrate repsettings command (#363)
* migrate repsettings command

* Update locale_es_ES.properties

* Update locale_pt_PT.properties
