+++ src/main/java/de/chojo/repbot/ReputationBot.java
@@ -19,8 +19,8 @@ import de.chojo.repbot.commands.messages.Messages;
 import de.chojo.repbot.commands.log.Log;
 import de.chojo.repbot.commands.repadmin.RepAdmin;
 import de.chojo.repbot.commands.reactions.Reactions;
-import de.chojo.repbot.commands.RepSettings;
 import de.chojo.repbot.commands.reputation.Reputation;
+import de.chojo.repbot.commands.repsettings.RepSettings;
 import de.chojo.repbot.commands.Roles;
 import de.chojo.repbot.commands.scan.Scan;
 import de.chojo.repbot.commands.Setup;
