+++ src/main/resources/locale.properties
@@ -150,54 +150,6 @@ command.reaction.remove.description=
 command.reaction.remove.emote.description=
 command.reaction.remove.message.notfound=
 command.reaction.remove.message.removed=
-command.repSettings.description=
-command.repSettings.embed.choose=
-command.repSettings.embed.descr.byAnswer=
-command.repSettings.embed.descr.byEmbed=
-command.repSettings.embed.descr.byFuzzy=
-command.repSettings.embed.descr.byMention=
-command.repSettings.embed.descr.byReaction=
-command.repSettings.embed.descr.emojidebug=
-command.repSettings.embed.descr.reputationMode=
-command.repSettings.embed.descr.skipSingleEmbed=
-command.repSettings.embed.title=
-command.repSettings.sub.answer.arg.answer=
-command.repSettings.sub.answer.false=
-command.repSettings.sub.answer.true=
-command.repSettings.sub.embed.arg.embed=
-command.repSettings.sub.embed.false=
-command.repSettings.sub.embed.true=
-command.repSettings.sub.emojidebug=
-command.repSettings.sub.emojidebug.arg.active=
-command.repSettings.sub.emojidebug.explain.cooldown=
-command.repSettings.sub.emojidebug.explain.donorLimit=
-command.repSettings.sub.emojidebug.explain.found=
-command.repSettings.sub.emojidebug.explain.noDonor=
-command.repSettings.sub.emojidebug.explain.noReceiver=
-command.repSettings.sub.emojidebug.explain.noRecentMessages=
-command.repSettings.sub.emojidebug.explain.prompted=
-command.repSettings.sub.emojidebug.explain.receiverLimit=
-command.repSettings.sub.emojidebug.explain.title=
-command.repSettings.sub.emojidebug.explain.tooOld=
-command.repSettings.sub.emojidebug.false=
-command.repSettings.sub.emojidebug.true=
-command.repSettings.sub.fuzzy.arg.fuzzy=
-command.repSettings.sub.fuzzy.false=
-command.repSettings.sub.fuzzy.true=
-command.repSettings.sub.info=
-command.repSettings.sub.mention.arg.mention=
-command.repSettings.sub.mention.false=
-command.repSettings.sub.mention.true=
-command.repSettings.sub.reactions.arg.reactions=
-command.repSettings.sub.reactions.false=
-command.repSettings.sub.reactions.true=
-command.repSettings.sub.reputationMode=
-command.repSettings.sub.reputationMode.rollingMonth=
-command.repSettings.sub.reputationMode.rollingWeek=
-command.repSettings.sub.reputationMode.total=
-command.repSettings.sub.skipSingleEmbed.arg.active=
-command.repSettings.sub.skipSingleEmbed.false=
-command.repSettings.sub.skipSingleEmbed.true=
 command.repadmin.description=
 command.repadmin.profile.description=
 command.repadmin.profile.user.description=
@@ -214,7 +166,54 @@ command.repadmin.reputation.set.amount.description=
 command.repadmin.reputation.set.description=
 command.repadmin.reputation.set.message.set=
 command.repadmin.reputation.set.user.description=
-command.repadmin.sub.reputation.arg.user=
+command.repsettings.description=
+command.repsettings.emojidebug.description=
+command.repsettings.emojidebug.message.cooldown=
+command.repsettings.emojidebug.message.donorLimit=
+command.repsettings.emojidebug.message.false=
+command.repsettings.emojidebug.message.found=
+command.repsettings.emojidebug.message.noDonor=
+command.repsettings.emojidebug.message.noReceiver=
+command.repsettings.emojidebug.message.noRecentMessages=
+command.repsettings.emojidebug.message.prompted=
+command.repsettings.emojidebug.message.receiverLimit=
+command.repsettings.emojidebug.message.title=
+command.repsettings.emojidebug.message.tooold=
+command.repsettings.emojidebug.message.true=
+command.repsettings.info.description=
+command.repsettings.info.message.answer.false=
+command.repsettings.info.message.answer.true=
+command.repsettings.info.message.choose=
+command.repsettings.info.message.embed.false=
+command.repsettings.info.message.embed.true=
+command.repsettings.info.message.fuzzy.false=
+command.repsettings.info.message.fuzzy.true=
+command.repsettings.info.message.mention.false=
+command.repsettings.info.message.mention.true=
+command.repsettings.info.message.option.byanswer.description=
+command.repsettings.info.message.option.byanswer.name=
+command.repsettings.info.message.option.byembed.description=
+command.repsettings.info.message.option.byembed.name=
+command.repsettings.info.message.option.byfuzzy.description=
+command.repsettings.info.message.option.byfuzzy.name=
+command.repsettings.info.message.option.bymention.description=
+command.repsettings.info.message.option.bymention.name=
+command.repsettings.info.message.option.byreaction.description=
+command.repsettings.info.message.option.byreaction.name=
+command.repsettings.info.message.option.emojidebug.description=
+command.repsettings.info.message.option.emojidebug.name=
+command.repsettings.info.message.option.reputationmode.description=
+command.repsettings.info.message.option.reputationmode.name=
+command.repsettings.info.message.option.skipsingleembed.description=
+command.repsettings.info.message.option.skipsingleembed.name=
+command.repsettings.info.message.reactions.false=
+command.repsettings.info.message.reactions.true=
+command.repsettings.info.message.reputationMode.rollingMonth=
+command.repsettings.info.message.reputationMode.rollingWeek=
+command.repsettings.info.message.reputationMode.total=
+command.repsettings.info.message.skipsingleembed.false=
+command.repsettings.info.message.skipsingleembed.true=
+command.repsettings.info.message.title=
 command.reputation.description=
 command.reputation.description.arg.mode=
 command.reputation.user.description=
